+++ tests/fixtures/52-multiline-import-with-comments/input.py
@@ -1,0 +1,5 @@
+from brownie import (
+    accounts,
+    network,  # for show_active
+    Contract,
+)