+++ tests/fixtures/53-multiple-from-brownie-lines/expected.py
@@ -1,0 +1,7 @@
+from ape import accounts
+from ape import networks
+from ape import chain
+
+
+def setup():
+    return accounts[0]