+++ tests/fixtures/53-multiple-from-brownie-lines/input.py
@@ -1,0 +1,7 @@
+from brownie import accounts
+from brownie import network
+from brownie import chain
+
+
+def setup():
+    return accounts[0]